Remove tags, use single string "paths" instead.
Rationale for using a single, long bytestring instead of multiple tags:

- It simplifies the API signficiantly, as you just need a pointer +
length instead of a pointer to a list of pointers + lengths... which may
look complicated if it is optimized.
- It is more flexible as it can work with both tag-based filesystems as
well as directory-based systems (à la UNIX).
- It is obviously faster to process.

Processes are still free to interpret the "path" in any way they like,
so not much changes there. it may be worth looking into having per-table
meta attributes that hint at the format of the paths.