Directly use physical addresses in virtio*, fix virtio-net driver
While using virtual addresses is more convienent, it is slower and
actually harder to get right as structures laid over a page boundary may
not be contiguous. It does make the API strictly unsafe, though this was
already pretty much the case. Making the API unsafe also allows us to pass
the responsibility of ensuring correct lifetime bounds to the user.

The virtio-net driver was broken since ´std::os::norostb::take_job` is a
blocking call. This is only fixable by creating a queue that is only
used by the driver. Currently, a thread is still necessary to regularly
poll the queue since `process_io_queue` is still blocking.