Implement proper userspace allocator, add syscall::{alloc, dealloc}
Currently, a whole page is allocated for every allocation, even if it's
very small. This is not efficient but currently the easiest option
(slabmalloc seems to be broken somehow, so that's not an option :( ).

The alloc and dealloc APIs are mostly page-agnostic. This is to give the
kernel more flexibility in how it allocates pages. It is still a
somewhat leaky abstraction as manually specified base addresses to alloc
must be aligned.

While userspace applications should no longer leak memory, the kernel
still does. This will be fixed in a later commit.