Implement process exit & cleanup
Thread waiting & exit hasn't been fully implemented yet but it should be
trivial.

Note that we don't check if a thread has been stopped. This isn't
strictly necessary right now since the kernel can only use a single core
anyways, but it will be a problem once we implement SMP.