Processs table, driver table, io::share
This is part of the work to implement `jail`. It is a lot more work
than anticipated, so a commit is being made now.

A process table has been added. This doesn't show all active processes
yet but it can be used to create a new process.

The driver table lists all drivers passed by the bootloader as files.
These will be used later by a single `init` process.

To pass objects to a newly created process, io::share has been added.
This is also used to specify the ELF binary in an efficient way.