Fix resuming threads from (timer) interrupts
I noticed that IRQ 16 was mapped to a noop handler for seemingly no
reason. After removing it, a #GP occured. It turns out that timer
interrupts got remapped to IRQ 16 by the scheduler.

After fixing that the kernel stack seemingly got corrupted as iretq
would frequently cause a #PF or #GP. After many hypotheses and fixes the
conclusion is that nesting IRQs is a bad idea, especially the timer IRQ.

By keeping the timer IRQ fixed the kernel_stack_ptr kept getting bumped
up every time it occured. However, as soon as a thread could be yielded
there was a high chance its the same thread the scheduler is using. The
scheduler in the meantime corrupted this stack, preventing iretq from
functioning.

The simple fix is to always yield to a "sleep" thread to avoid having to
nest IRQs. This works wonderfully well and I recommend it to any OS
developers reading this :)

Some of the other fixes that have been made are:

- compile_error! in the PIC and RTC drivers since they haven't been
  tested at all.
- disabled the rtc driver entirely
- Moved the PIC initialization code to arch::amd64 just in case.
- Mask IF when performing syscall