Fix more mutex locks while interrupts are disabled
I have given this some thought and there are two main issues:

- During init interrupts are disabled right until the very end.
- Interrupts are disabled while destroying a thread (and process, by
  extension).

The former can be fixed by creating a kernel thread to handle the
remainder of the init phase. The latter can be fixed by either creating
a kernel thread for handling cleanup or somehow "demoting" the thread
that is being destroyed to a kernel thread first, alleviating the need
to switch to the (interrupt-free) CPU-local stack.