Fix race condition with Thread::sleep
Consider the following situation:

    poll Futures

    check if any futures are ready

					complete future, wake thread

    sleep if no futures are ready

Waking a thread sets sleep to 0, sleeping sets it to infinity. This led
to a situation where a thread effectively did not get woken up.

This has been fixed by adding a flag that is set along with sleep
atomically when a thread is woken up. If the thread tries to sleep but
this flag is set it clears it and returns immediately.