Fix potential race condition in halt_forever() thread
If an interrupt which wakes a thread happens right before `hlt` the
halt_forever() task would continue instead of the woken thread.

`sti; hlt` is atomic, i.e. a pending interrupt won't be halted until
`hlt` is executed. The one exception is NMIs, but we can check in the
handler for the `sti; hlt` sequence and step over `hlt` before
returning.