Add portio driver
As it turns out, it is _not_ always possible to access the VGA registers
from MMIO space. To keep legacy compatibility cruft to a minimum, a Port
I/O object can be used to write to the entire I/O space. This is wildly
unsafe but by far the simplest method.