Make the heap allocator page-aware
While the "page-agnostic" alloc/dealloc was a nice idea, in practice it
is tricky to implement correctly. A performant heap allocator needs to
be aware of page sizes anyways to avoid redundant syscalls anyways, so
there was little benefit to it after all.

Also fix some more alignment bugs (sigh)