Significantly optimize rendering
Rendering a single character went from ~30ms to ~2ms, which is still bad
but a ~15x speedup regardless.

There are quite a few changes that led to this improvement:

- MemoryRange has been removed in favor of SubRange. MemoryRange is too
  hard to get right with multiple objects and gaps in memory whereas Subrange
  only wraps a range of a MemoryObject. It is also created with a handle
  instead of a address range, which alleviates the need to deal with
  gaps.
- A SharedMemory object has been added for userspace. This is a type of
  memory that is explictly intended for sharing with other processes.
  Memory allocated with alloc() is intended to be private and
  unshareable.
- StreamTable can handle share operations. An object is (temporarily)
  held in the table when shared and the server removes the object by
  opening it.

SharedMemory allows copying data directly from one process to another
and only transfer a tiny amount of data through the kernel when sending
notifications. It also prevents the need to constantly map/unmap pages,
reducing TLB pressure.

One obvious limitation is that SharedMemory only works on a single
machine and not between machines. This probably isn't a problem though,
since a process on one machine can read & serialize the data and transfer
it to a process on another machine which then deserializes it.