Replace dma_alloc syscall with table, move post_init to kernel thread
I intended to do this in two commits but I got sidetracked.

DMA allocation is reserved for drivers and other applications should not
need it. The proper way to restrict it is by moving the functionality to
an object, then removing access to that object.

A large chunk of the initialization code has been moved to a kernel
thread. This will allow reducing the amount of spinlocks signficantly.

Not all of the code has been updated to accomodate for these changes as
none of it is final yet.