Add do_io syscall for synchronous requests
Synchronous requests are very common in simple programs. To simplify &
reduce the size of these programs a system call has been added which can
be used to send one I/O request at a time. It should also be a bit more
performant as there is no need redundant copies & poll + wait as with
the asynchronous I/O queue.