async_std basics
This async_std is completely unrelated to the one on crates.io.
This version is specifically created for Norost B and makes heavy use of the
async I/O queue.

scancode_to_char has been adapted to make use of it.

tokio_uring has been used as inspiration for the API and improvements to
the I/O queue.
For example, the queue no longer directly takes ownership of the buffers
and instead relies on the BufferFuture to keep it alive, which allows
using containers other than Vec.