Implement message pipes
They're simimal to regular pipes but work with messages that have a
well-defined size instead of a contiguous data stream.

While this could have been hacked on top of the regular data stream in
userspace it is signficantly easier to keep a clear message boundary
all the way from sender to receiver.