Add framebuffer driver
The framebuffer is provided by the bootloader, is always useable and
doesn't require writing GPU drivers.

There is some PAT stuff as firmware doesn't set MTRRs to WC for the
framebuffer but Intel didn't make it easy to use properly and *very*
easy to fuck up, so full support is going to be implemented in a
separate commit.